+++ src/r_plane.c
@@ -53,6 +53,9 @@
 #include "w_wad.h"
 #include "z_zone.h"
 
+// [Nugget]
+#include "i_video.h"
+
 #define MAXVISPLANES 128    /* must be a power of 2 */
 
 static visplane_t *visplanes[MAXVISPLANES];   // killough
@@ -406,7 +409,7 @@ static void do_draw_plane(visplane_t *pl)
 	angle_t an, flip;
 	boolean vertically_scrolling = false;
 	boolean stretch;
-	int skyheight_target; // [Nugget] Stretch sky just as much as necessary
+	int skyheight_target = 0; // [Nugget] Stretch sky just as much as necessary
 
 	// killough 10/98: allow skies to come from sidedefs.
 	// Allows scrolling and/or animated skies, as well as
@@ -467,11 +470,21 @@ static void do_draw_plane(visplane_t *pl)
         dc_iscale = skyiscale;
 
         // [FG] stretch short skies
-        
-        // [Nugget] Stretch sky just as much as necessary
-        skyheight_target = 200 - (dc_texturemid >> FRACBITS);
-        
-        stretch = (stretchsky && dc_texheight < skyheight_target);
+
+        // [Nugget] /---------------------------------------------------------
+
+        // Stretch sky just as much as necessary
+        skyheight_target = (stretchsky ? 200 : 100) - (dc_texturemid >> FRACBITS);
+
+        // FOV-based sky stretching
+        if (fov_stretchsky && skyiscalediff > FRACUNIT)
+        {
+          skyheight_target = skyheight_target * skyiscalediff / FRACUNIT;
+        }
+
+        // [Nugget] ---------------------------------------------------------/
+
+        stretch = (dc_texheight < skyheight_target);
         if (stretch || !vertically_scrolling)
         {
           fixed_t diff;